webapp/requirements.txt
flask==3.0.0
lxml==4.9.3
python-dotenv==1.0.0
